menambahkan fungsi post dan put

+++ src/main/java/buildrestful/restful1/ProductServiceController.java
@@ -44,23 +44,34 @@ public class ProductServiceController {
    public ResponseEntity<Object> getProduct() {
       return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
    }
-    @RequestMapping(value = "/products", method = RequestMethod.POST)
-   public ResponseEntity<Object> createProduct(@RequestBody Product product) {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+   @RequestMapping(value = "/products", method = RequestMethod.POST)
+        public ResponseEntity<Object> createProduct(@RequestBody Product product){
+        //jika ada id sama akan memunculkan "Product key cannot duplicated"
+        if (productRepo.containsKey(product.getId())){
+           return new ResponseEntity<>("Produk tidak boleh sama", HttpStatus.OK);
+        }
+        else{
+           productRepo.put(product.getId(), product);
+           return new ResponseEntity<>("Produk sudah dibuat", HttpStatus.CREATED);
+        }
    }
     @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable("id") String id){
         productRepo.remove(id);
-        return new ResponseEntity<>("Product is deleted succsfully", HttpStatus.OK);      
+        return new ResponseEntity<>("produk sudah dihapus", HttpStatus.OK);      
     }
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
-   public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+        public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
+        if(!productRepo.containsKey(id)){ 
+           return new ResponseEntity<>("produk key belum ada ", HttpStatus.NOT_FOUND);
+        }
+        else{
+           productRepo.remove(id);
+           product.setId(id);
+           productRepo.put(id, product);
+           return new ResponseEntity<>("Produk sudah di Update", HttpStatus.OK);
     }
+        }
 }
